Refactor delete user form
- use modal components
- seperate from logic and display logic
